refactor : #35 -UserResponse/UserJoinResponse 통합
