+++ src/test/java/com/couple/sns/domain/post/controller/CommentControllerTest.java
@@ -14,8 +14,8 @@ import com.couple.sns.common.exception.ErrorCode;
 import com.couple.sns.common.exception.SnsApplicationException;
 import com.couple.sns.domain.post.dto.response.CommentResponse;
 import com.couple.sns.domain.post.dto.response.LikeResponse;
-import com.couple.sns.domain.post.dto.response.UserResponse;
 import com.couple.sns.domain.post.service.CommentService;
+import com.couple.sns.domain.user.dto.response.UserResponse;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import java.util.ArrayList;
 import java.util.List;
