+++ src/test/java/com/couple/sns/domain/post/controller/PostControllerTest.java
@@ -18,8 +18,8 @@ import com.couple.sns.domain.post.dto.PostDto;
 import com.couple.sns.domain.post.dto.request.PostRequest;
 import com.couple.sns.domain.post.dto.response.LikeResponse;
 import com.couple.sns.domain.post.dto.response.PostResponse;
-import com.couple.sns.domain.post.dto.response.UserResponse;
 import com.couple.sns.domain.post.service.PostService;
+import com.couple.sns.domain.user.dto.response.UserResponse;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import java.util.ArrayList;
 import java.util.List;
